pyproject.toml
[tool.poetry]
name = "johndoe"
version = "0.1.0"
description = ""
authors = ["sturdy-robot <pedrenriquegg@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pygame-ce = "^2.5.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
pyinstaller = "^6.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
